Move npm dependencies module to lib folder
