+++ src/lib/npm-dependencies.js
